refactor(bee-keeper): move flat src into aura_keeper package
Second app in the namespace refactor (after mcp-server #224). bee-keeper's
code lived as a root main.py plus flat src/ modules (config) + a generic
src/hive/ package — the flat modules collided with other services' main/config
and hive when installed editable.

- Move main.py (from the project root), src/config.py, src/hive/ and
  src/__init__.py into src/aura_keeper/ (git-tracked renames).
- Rewrite the 7 intra-app imports to the aura_keeper namespace
  (from config -> from aura_keeper.config; from hive.X -> from aura_keeper.hive.X).
- pyproject: entry point -> aura_keeper.main:main; switch the implicit
  setuptools build to hatchling with packages = ["src/aura_keeper"]; drop the
  stale aura_keeper.egg-info artifact.
- Fix a pre-existing bug the move exposed to mypy: main.py called
  metabolism.execute(signal=..., event_name=...) but execute() only accepts
  event_name (would TypeError at runtime; the arg was never used). main.py sat
  at the repo root, outside the mypy-checked src/ tree, so it was never linted.
- Update the (skipped) run command comment for the new module path.

Also add fastmcp.* to the root mypy ignore-missing-imports overrides: it was
missing from the list of stub-less third-party libs, so `make lint` fails on
synapses/mcp-server in a clean env (no fastmcp installed). This makes the
mcp-server mypy pass deterministic, matching how litellm/dspy/nats/etc. are
handled.

Verified: `make lint` green (all services incl. bee-keeper 15 files);
`uv build --package aura-keeper --wheel` ships only aura_keeper/; full import
chain (main -> hive.metabolism -> config) imports cleanly.
