+++ agents/bee-keeper/pyproject.toml
@@ -19,7 +19,16 @@ dependencies = [
 aura-core = { workspace = true }
 
 [project.scripts]
-aura-keeper = "main:main"
+aura-keeper = "aura_keeper.main:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+# Standard src-layout: ship the aura_keeper package (no flat top-level modules,
+# no `hive`/`config` collision with other services).
+[tool.hatch.build.targets.wheel]
+packages = ["src/aura_keeper"]
 
 [tool.uv]
 managed = true
