+++ agents/bee-keeper/src/aura_keeper/hive/connector/proteins/vcs/manifest.yaml
