+++ agents/bee-keeper/src/aura_keeper/hive/connector/proteins/vcs/schema.py
