+++ agents/bee-keeper/src/aura_keeper/hive/connector/proteins/vcs/skill.py
