+++ agents/bee-keeper/src/aura_keeper/hive/metabolism.py
@@ -2,7 +2,7 @@ import time
 
 import structlog
 
-from config import KeeperSettings
+from aura_keeper.config import KeeperSettings
 from aura_core_gen.aura.core.v1 import (
     AuditObservation,
 )
