+++ agents/bee-keeper/src/aura_keeper/hive/transformer/__init__.py
@@ -6,7 +6,7 @@ import litellm
 import structlog
 import yaml  # type: ignore
 
-from config import KeeperSettings
+from aura_keeper.config import KeeperSettings
 from aura_core import (
     ALLOWED_CHAMBERS,
     ALLOWED_ROOT_FILES,
