+++ agents/bee-keeper/src/aura_keeper/hive/transformer/signatures.py
